[CI] Relocate server test cases from ci_use directory to e2e
