+++ tests/e2e/EB_VL_Lite/test_rollout_model.py
